fixes pr
