+++ Dispatcher/UI/NavigationBar.swift
@@ -8,7 +8,6 @@ protocol HeaderDelegate: AnyObject {
 class NavigationBar {
     weak var delegate: HeaderDelegate?
     
-    
     func setupNavigationBar(for viewController: UIViewController) {
         let iconImage = UIImage(named: "logo")
         let iconImageView = UIImageView(image: iconImage)
@@ -20,8 +19,6 @@ class NavigationBar {
         
         viewController.navigationItem.rightBarButtonItems = [alertButton, searchButton]
         viewController.navigationItem.leftBarButtonItems = [iconBarButton]
-        
-    
     }
     
     @objc private func alertButtonTapped() {
