--- Dispatcher/UI/Screens/FavoritePage/FavoritePageViewModel.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-class FavoritePageViewModel {}
