+++ Dispatcher/UI/Screens/SplashPage/SplashViewController.swift
@@ -2,19 +2,21 @@ import UIKit
 
 class SplashViewController: UIViewController {
     
-    static var tabBarIdentifier = "TabBarController"
-    static var storyboardName = "Main"
-
     override func viewDidLoad() {
         super.viewDidLoad()
     }
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-
+        delaySplash()
+    }
+    
+    func delaySplash() {
+        let tabBarIdentifier = "TabBarController"
+        let storyboardName = "Main"
         DispatchQueue.main.asyncAfter(deadline: .now() + 3.0) {
-            let storyboard = UIStoryboard(name: SplashViewController.storyboardName, bundle: nil)
-            let tabBarController = storyboard.instantiateViewController(withIdentifier: SplashViewController.tabBarIdentifier) as! UITabBarController
+            let storyboard = UIStoryboard(name: storyboardName, bundle: nil)
+            let tabBarController = storyboard.instantiateViewController(withIdentifier: tabBarIdentifier) as! UITabBarController
             tabBarController.modalPresentationStyle = .fullScreen
             self.present(tabBarController, animated: false, completion: nil)
         }
