--- Dispatcher/UI/Screens/UserPage/UserPageViewModel.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-class UserPageViewModel {}
-
